Role comoplete

+++ app/Http/Controllers/RoleController.php
@@ -4,13 +4,15 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Role;
+use App\Models\School;
 
 class RoleController extends Controller
 {
     public function index()
     {
-        $roles = Role::all();
-        return view('role-list',compact('roles'));
+        $schools = School::all();
+        $roles = Role::with('school')->get();
+        return view('role-list',compact('roles','schools'));
     }
     
     public function getRole($id)
@@ -21,6 +23,9 @@ class RoleController extends Controller
     
     public function deleteRole($id)
     {
+        if($id == 1 || $id == 2){
+            return back()->with('error','Admin Role is not delete.');
+        }
         $role = Role::where('id',$id)->first();
         $role->delete();
         return back()->with('success','Delete Role Successfully.');
@@ -29,13 +34,13 @@ class RoleController extends Controller
     public function saveRole(Request $request)
     {
         $request->validate([
-            'school_name' => 'required|string',
+            'school_id' => 'required|integer|exists:schools,id',
             'role_name' => 'required|string',
         ]);
 
         Role::create([
             'role_name' => $request->role_name,
-            'school_id' => 1
+            'school_id' => $request->school_id
         ]);
         
         return back()->with('success','Role Create Successfully');
@@ -45,8 +50,8 @@ class RoleController extends Controller
     {
 
         $request->validate([
-            'school_name' => 'required|string',
-            'id' => 'required|integer',
+            'school_id' => 'required|integer|exists:schools,id',
+            'id' => 'required|integer|exists:roles,id',
             'role_name' => 'required|string',
         ]);
 
@@ -56,7 +61,7 @@ class RoleController extends Controller
         }
         
         $role->role_name = $request->role_name;
-        $role->school_id = $request->school_name;
+        $role->school_id = $request->school_id;
         $role->save();        
         
         return back()->with('success','Role Save Successfully');
